CTK used from the cim-unito fork on GitHub
The revision tag is not changed, only the repository. It means that
we always will rely on the same version of CTK that the official MITK
relies on. However, we use the CIM fork, where our patches are merged.

On the other side, this also means, that if MITK is updated to use a
newer version of CTK that is not merged to the CIM fork then MITK
(and so KMaps) will not compile.

Then you have to do the following:

Clone the CIM fork of CTK, or step into a directory that already
contains a clone (e.g. MITK-superbuild/CMakeExternals/Sources/CTK).

$ git checkout cim
$ git pull origin cim

If you do not have an "upstream" alias yet:
$ git remote add upstream git://github.com/commontk/CTK.git

$ git fetch upstream
$ git merge upstream/master

Resolve conflicts, if any.

$ git push origin cim